fix: Corregi unos detalles para ver como debemos cambiar la vista de los comandos
