+++ .eslintrc.json
@@ -18,7 +18,7 @@
     "react/prop-types": "off",
     "eqeqeq": "warn",
     "semi": ["off", "never"],
-    "indent": ["error", 2],
-    "quotes": ["error", "single"]
+    "indent": ["warn", 2],
+    "quotes": ["warn", "single"]
   }
 }
