--- src/components/Commands/CommandContent/CommandContent.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function CommandContent({ description }){
-  return (
-    <div className="CommandContent-Container">
-      <p className="CommandContent-Container__p">{ description }</p>
-    </div>
-  )
-}
-
