--- src/components/Commands/CommandContent/CommandContent.scss
@@ -1,7 +1,0 @@
-.CommandContent-Container {
-  margin: 1rem 0 0 2.5rem;
-
-  &__p {
-    font-family: var(--lato-font);
-  }
-}
