+++ src/components/CommandsContainer/CommandsContainer.scss
@@ -1,59 +1,38 @@
-.CommandsContainer {
-  &__content {
-    &__title {
-      font-size: 1.5rem;
-      font-family: var(--roboto-mono-font);
+.CommandContainer {
+  @include layout(20px);
+
+  &__command__name {
+    display: inline-block;
+    align-items: center;
+    font-size: 1.5rem;
+    color: var(--orange-principal);
+    font-family: var(--roboto-mono-font);
+    font-weight: bold;
+    margin: 1rem 0 1rem 0;
+    outline: none;
+    cursor: pointer;
+    opacity: 0.85;
+    transition: var(--transition);
+
+    i {
+      color: var(--orange-sec);
+      opacity: 0;
+      visibility: hidden;
+      margin-right: 0.5rem;
     }
 
-    &__flex {
-      align-items: center;
-      column-gap: 1.4rem;
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: flex-start;
-      margin-top: 1.6rem;
-      margin-bottom: 3rem;
-    }
+    &:hover {
+      opacity: 1;
+      transition: var(--transition);
 
-    &-btnOs {
-      @include btnStyle;
-      color: var(--black-principal);
-      background: var(--orange-sec);
-      border: 2px solid var(--black-sec);
-
-      &:nth-child(1) {
-        width: 100px;
+      i {
+        opacity: 1;
+        visibility: visible;
       }
     }
 
-    &-btnConfig {
-      @include btnStyle(110px);
-      border: 3px solid var(--blue-sec);
-      background: var(--blue-principal);
-      color: var(--white-principal);
-      margin-top: 1.3rem;
-    }
-
-    &-btnRepo {
-      @include btnStyle;
-      color: var(--white-principal);
-      background: var(--blue-principal);
-      border: 3px solid var(--blue-sec);
-    }
-
-    &-btnKnowMore {
-      @include btnStyle(auto);
-      border: 2px solid var(--orange-principal);
-      background: var(--white-sec);
-      color: var(--orange-principal);
-      margin-bottom: 1rem;
-      padding: 0.75rem 1rem;
-
-      @media (min-width: $ipad) {
-        &:hover {
-          box-shadow: 1.1px 1.3px 3px var(--shadow);
-        }
-      }
+    span {
+      font-family: var(--lato-font);
     }
   }
 }
