+++ src/components/MenuMobile/MenuContent/MenuContent.scss
@@ -1,0 +1,59 @@
+.CommandsContainer {
+  &__content {
+    &__title {
+      font-size: 1.5rem;
+      font-family: var(--roboto-mono-font);
+    }
+
+    &__flex {
+      align-items: center;
+      column-gap: 1.4rem;
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: flex-start;
+      margin-top: 1.6rem;
+      margin-bottom: 3rem;
+    }
+
+    &-btnOs {
+      @include btnStyle;
+      color: var(--black-principal);
+      background: var(--orange-sec);
+      border: 2px solid var(--black-sec);
+
+      &:nth-child(1) {
+        width: 100px;
+      }
+    }
+
+    &-btnConfig {
+      @include btnStyle(110px);
+      border: 3px solid var(--blue-sec);
+      background: var(--blue-principal);
+      color: var(--white-principal);
+      margin-top: 1.3rem;
+    }
+
+    &-btnRepo {
+      @include btnStyle;
+      color: var(--white-principal);
+      background: var(--blue-principal);
+      border: 3px solid var(--blue-sec);
+    }
+
+    &-btnKnowMore {
+      @include btnStyle(auto);
+      border: 2px solid var(--orange-principal);
+      background: var(--white-sec);
+      color: var(--orange-principal);
+      margin-bottom: 1rem;
+      padding: 0.75rem 1rem;
+
+      @media (min-width: $ipad) {
+        &:hover {
+          box-shadow: 1.1px 1.3px 3px var(--shadow);
+        }
+      }
+    }
+  }
+}
