+++ src/components/NavMobile/NavMobile.scss
@@ -4,7 +4,6 @@
   align-content: center;
   display: flex;
   justify-content: space-between;
-  padding-top: 20px;
 
   img {
     @include userNoSelect;
