+++ src/components/_index.scss
@@ -2,9 +2,8 @@
 @import './ContentHome/ContentHome.scss';
 @import './NavMobile/NavMobile.scss';
 @import './MenuMobile/MenuMobile.scss';
-@import './CommandsContainer/CommandsContainer.scss';
+@import './MenuMobile/MenuContent/MenuContent.scss';
 @import './InstallWindows/installWindows.scss';
 @import './BtnToTop/BtnToTop.scss';
 @import './GitConfigView/GitConfigView.scss';
-@import '../components/Commands/Commands.scss';
-@import '../components/Commands/CommandContent/CommandContent.scss';
+@import '../components/CommandsContainer/CommandsContainer.scss';
