+++ src/index.scss
@@ -19,5 +19,3 @@
 /* SCSS Components Files */
 @import './views/Layout/Layout.scss';
 @import './components/index.scss';
-@import '../src/components/Commands/Commands.scss';
-@import '../src/components/Commands/CommandContent/CommandContent.scss';
