+++ src/scss/mixins/_linkStyle.scss
@@ -1,3 +1,8 @@
+/*
+* Esto solo le da un estilo a los estilos
+* para que combinen con la pagina
+*
+*/
 @mixin linkStyle {
   display: block;
   color: var(--blue-sec);
