+++ src/views/Layout/Layout.jsx
@@ -8,8 +8,7 @@ import MenuMobile from '../../components/MenuMobile/MenuMobile';
 import InstallWindows from '../../components/InstallWindows/InstallWindows';
 import BtnToTop from '../../components/BtnToTop/BtnToTop';
 import GitConfigView from '../../components/GitConfigView/GitConfigView';
-import Commands from '../../components/Commands/Commands';
-import CommandContent from '../../components/Commands/CommandContent/CommandContent';
+import CommandsContainer from '../../components/CommandsContainer/CommandsContainer';
 
 export default function Layout() {
   const [showMenu, setShowMenu] = useState(false);
@@ -40,11 +39,11 @@ export default function Layout() {
             <Route path='/config'>
               <GitConfigView />
             </Route>
-            <Route path="/commands">
-              <Commands/>
+            <Route path='/commands'>
+              <CommandsContainer />
             </Route>
-            <Route path="/git-init">
-              <CommandContent description="Prueba desde layout"/>
+            <Route path='/commands/git-init'>
+              <ContentComponent />
             </Route>
           </Switch>
         )}
@@ -53,4 +52,3 @@ export default function Layout() {
     </Router>
   );
 }
-
